+++ .simple-git-hooks.js
@@ -1,3 +1,0 @@
-module.exports = {
-  'pre-commit': 'npx lint-staged',
-};